refactor(rsc): clarify cache action filenames
