+++ packages/plugin-rsc/examples/use-cache-persistent/src/features/file-directive-from-server/action-cached.ts
@@ -1,0 +1,9 @@
+'use cache'
+
+import { getDirectLabel } from './direct'
+import { state } from './state'
+
+export async function cachedFromServer(argument: string) {
+  state.executionCount++
+  return `server import + body-v1 + ${getDirectLabel()} + ${argument}`
+}
