--- packages/plugin-rsc/examples/use-cache-persistent/src/features/file-directive-from-server/reset.ts
@@ -1,9 +1,0 @@
-'use server'
-
-import { resetCache } from '../../framework/use-cache-runtime'
-import { state } from './state'
-
-export async function resetAction() {
-  await resetCache()
-  state.executionCount = 0
-}
